junk/embedder/pyproject.toml
[project]
name = "embedder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.1",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.0",
    "pyarrow>=21.0.0",
    "rich>=14.1.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "typer>=0.19.2",
    "umap-learn>=0.5.9.post2",
]
